+++ SynchronisedMetadata.cs
@@ -15,71 +15,14 @@ namespace ContentWarningShop
     /// <typeparam name="T"></typeparam>
     public class SynchronisedMetadata<T> where T : IConvertible, IComparable
     {
-        #region static
-        private static Callback<LobbyChatUpdate_t> cb_onLobbyStatusUpdate;
-        private static Callback<LobbyCreated_t> cb_onLobbyCreated;
-        private static Callback<LobbyEnter_t> cb_onLobbyEntered;
-        private static event Action? _onLobbyJoined;
-        private static Callback<LobbyDataUpdate_t> cb_onLobbyDataUpdate;
-        private static event Action? _onLobbyDataUpdate;
-
-        private static CSteamID _currentLobby;
-        public static bool InLobby => _currentLobby != default;
-        public static bool IsHost => InLobby && SteamMatchmaking.GetLobbyOwner(_currentLobby) == SteamUser.GetSteamID();
-        
-
-        static SynchronisedMetadata()
-        {
-            cb_onLobbyStatusUpdate = Callback<LobbyChatUpdate_t>.Create(Steam_LobbyLeft);
-            cb_onLobbyCreated = Callback<LobbyCreated_t>.Create(Steam_LobbyCreated);
-            cb_onLobbyEntered = Callback<LobbyEnter_t>.Create(Steam_LobbyEntered);
-            cb_onLobbyDataUpdate = Callback<LobbyDataUpdate_t>.Create(Steam_LobbyDataUpdated);
-        }
-
-        private static void Steam_LobbyCreated(LobbyCreated_t e)
-        {
-            if (e.m_eResult == EResult.k_EResultOK)
-            {
-                _currentLobby = new CSteamID(e.m_ulSteamIDLobby);
-                _onLobbyJoined?.Invoke();
-            }
-        }
-
-        private static void Steam_LobbyEntered(LobbyEnter_t e)
-        {
-            // If we created the lobby, don't call the event again. Otherwise _currentLobby will be default by now
-            if (_currentLobby == default)
-            {
-                _currentLobby = new CSteamID(e.m_ulSteamIDLobby);
-                _onLobbyJoined?.Invoke();
-            }
-        }
-
-        private static void Steam_LobbyLeft(LobbyChatUpdate_t e)
-        {
-            var user = new CSteamID(e.m_ulSteamIDUserChanged);
-            if (user != SteamUser.GetSteamID())
-            {
-                return;
-            }
-            if ((EChatMemberStateChange)e.m_rgfChatMemberStateChange != EChatMemberStateChange.k_EChatMemberStateChangeEntered)
-            {
-                // Lobby left
-                _currentLobby = default;
-            }
-        }
-
-        private static void Steam_LobbyDataUpdated(LobbyDataUpdate_t e)
-        {
-            if (InLobby == false || e.m_ulSteamIDLobby != e.m_ulSteamIDMember)
-            {
-                return;
-            }
-            _onLobbyDataUpdate?.Invoke();
-        }
-        #endregion
-
-        #region instance
+        /// <summary>
+        /// Get if the player is the host of the current lobby. Will be <see langword="false"/> if <see cref="InLobby"/> is.
+        /// </summary>
+        public bool IsHost => SteamLobbyMetadataHandler.IsHost;
+        /// <summary>
+        /// Get if the player is in a Steam Lobby.
+        /// </summary>
+        public bool InLobby => SteamLobbyMetadataHandler.InLobby;
         /// <summary>
         /// The Steam Lobby Metadata key this instance is bound to.
         /// </summary>
@@ -96,8 +39,11 @@ namespace ContentWarningShop
             get => _value;
         }
         /// <summary>
-        /// Checks if the entry is being actively synced with a lobby.
+        /// Checks if the entry is being actively synced with a lobby. Is <see langword="false"/> if the player is not currently in a lobby.
         /// </summary>
+        /// <remarks>
+        /// To check if this instance will send and receive updates from the SteamAPI, see <see cref="IsConnected"/> instead.
+        /// </remarks>
         public bool IsSynced
         {
             get
@@ -106,7 +52,7 @@ namespace ContentWarningShop
                 {
                     return false;
                 }
-                return _currentLobby == default;
+                return SteamLobbyMetadataHandler.InLobby;
             }
         }
         /// <summary>
@@ -138,14 +84,14 @@ namespace ContentWarningShop
             {
                 throw new ArgumentException($"Key length must not exceed {nameof(Steamworks.Constants.k_nMaxLobbyKeyLength)} ({Steamworks.Constants.k_nMaxLobbyKeyLength}); was {key.Length}", nameof(key));
             }
-            _onLobbyDataUpdate += OnLobbyUpdate;
-            _onLobbyJoined += OnLobbyJoin;
+            SteamLobbyMetadataHandler.OnLobbyDataUpdate += OnLobbyUpdate;
+            SteamLobbyMetadataHandler.OnLobbyJoined += OnLobbyJoin;
             Key = key;
             _value = initialValue;
             // If we are creating the instance late, check if the key already has a registered value or not. If not, create it, if yes, fetch it.
-            if (InLobby)
+            if (SteamLobbyMetadataHandler.InLobby)
             {
-                var valStr = SteamMatchmaking.GetLobbyData(_currentLobby, Key);
+                var valStr = SteamMatchmaking.GetLobbyData(SteamLobbyMetadataHandler.CurrentLobby, Key);
                 if (string.IsNullOrEmpty(valStr))
                 {
                     SetValue(_value);
@@ -162,7 +108,7 @@ namespace ContentWarningShop
         /// Checks if assigning a new value is possible from this client.
         /// </summary>
         /// <remarks>
-        /// If the client is not in a lobby, this method returns true and the set value is treated as initialisation before joining one.
+        /// If the client is not in a lobby, returns true, otherwise checks if the player is the lobby's host.
         /// </remarks>
         /// <returns></returns>
         public bool CanSet()
@@ -173,6 +119,9 @@ namespace ContentWarningShop
         /// <summary>
         /// Sets the entry to a new value.
         /// </summary>
+        /// <remarks>
+        /// Setting a new value is only possible by the host of the lobby, or if the player in not in a lobby.
+        /// </remarks>
         /// <param name="value"></param>
         /// <returns>
         /// <see langword="true"/> if assigning the new value from this client was possible, <see langword="false"/> if not.
@@ -189,12 +138,15 @@ namespace ContentWarningShop
                 ValueChanged?.Invoke(_value);
                 return true;
             }
-            var canSet = CanSet();
-            if (canSet)
+            if (IsHost)
             {
-                SteamMatchmaking.SetLobbyData(_currentLobby, Key, ValToString(value));
+                if (SteamMatchmaking.SetLobbyData(SteamLobbyMetadataHandler.CurrentLobby, Key, ValToString(value)) == false)
+                {
+                    Debug.LogError($"Could not set {Key} to {_value}, despite being the lobby host.");
+                    return false;
+                }
             }
-            return canSet;
+            return true;
         }
 
         /// <summary>
@@ -204,17 +156,21 @@ namespace ContentWarningShop
         public void Disconnect()
         {
             IsConnected = false;
-            _onLobbyDataUpdate -= OnLobbyUpdate;
-            _onLobbyJoined -= OnLobbyJoin;
+            SteamLobbyMetadataHandler.OnLobbyDataUpdate -= OnLobbyUpdate;
+            SteamLobbyMetadataHandler.OnLobbyJoined -= OnLobbyJoin;
         }
 
         private void OnLobbyJoin()
         {
             if (IsHost)
             {
-                if (SteamMatchmaking.SetLobbyData(_currentLobby, Key, ValToString(_value)))
+                if (SteamMatchmaking.SetLobbyData(SteamLobbyMetadataHandler.CurrentLobby, Key, ValToString(_value)))
                 {
                     Debug.Log($"Set join lobby metadata {Key} to {_value} as host.");
+                }
+                else
+                {
+                    Debug.LogError($"Could not set {Key} to {_value}, despite being the lobby host.");
                 }
             }
             else
@@ -233,11 +189,11 @@ namespace ContentWarningShop
         /// </summary>
         protected void FetchValue()
         {
-            if (InLobby == false)
+            if (SteamLobbyMetadataHandler.InLobby == false)
             {
                 return;
             }
-            var valStr = SteamMatchmaking.GetLobbyData(_currentLobby, Key);
+            var valStr = SteamMatchmaking.GetLobbyData(SteamLobbyMetadataHandler.CurrentLobby, Key);
             if (string.IsNullOrEmpty(valStr))
             {
                 return;
@@ -260,6 +216,5 @@ namespace ContentWarningShop
         {
             return (TRes)Convert.ChangeType(value, typeof(TRes), CultureInfo.InvariantCulture);
         }
-        #endregion
     }
 }